search_app/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
prometheus-client==0.17.1

